refactor: split sparse filesystem weight sync into separate classes
Extract sparse weight broadcast into its own trainer class
(SparseFileSystemWeightBroadcast) and receiver class
(SparseFileSystemWeightUpdateWorker), removing all sparse logic from
the base FileSystemWeightBroadcast/Worker.

Config changes:
- New SparseFileSystemWeightBroadcastConfig with type='sparse_filesystem'
  and kernel_format field. Removes sparse/kernel_format from
  FileSystemWeightBroadcastConfig.
- Add 'sparse_filesystem' to SharedWeightBroadcastConfig,
  InferenceWeightBroadcastConfig, and OrchestratorFileSystemWeightBroadcastConfig
  type discriminators.

Trainer:
- FileSystemWeightBroadcast: clean, no sparse branches.
- SparseFileSystemWeightBroadcast (new): inherits from FileSystemWeightBroadcast,
  owns baseline capture, kernel/HF format collection, sparse patch writing.
- filter_state_dict_by_layers moved to trainer/weights.py (shared).

Receiver:
- FileSystemWeightUpdateWorker: clean, no sparse dispatch.
- SparseFileSystemWeightUpdateWorker (new): inherits from
  FileSystemWeightUpdateWorker, auto-detects kernel vs HF patches from
  manifest compute_dtype, dispatches to index_copy_ or CPU cache path.
- Registered in WORKER_EXTENSION_CLS as 'sparse_filesystem'.

All 7 tests pass (3 CPU + 4 GPU).
