+++ src/prime_rl/inference/vllm/server.py
@@ -72,6 +72,7 @@ def models(request: Request) -> OpenAIServingModels:
 WORKER_EXTENSION_CLS = {
     "nccl": "prime_rl.inference.vllm.worker.nccl.NCCLWeightUpdateWorker",
     "filesystem": "prime_rl.inference.vllm.worker.filesystem.FileSystemWeightUpdateWorker",
+    "sparse_filesystem": "prime_rl.inference.vllm.worker.sparse_filesystem.SparseFileSystemWeightUpdateWorker",
 }
