+++ src/prime_rl/trainer/rl/broadcast/__init__.py
@@ -6,6 +6,7 @@ from prime_rl.configs.trainer import LoRAConfig, WeightBroadcastConfig
 from prime_rl.trainer.rl.broadcast.base import WeightBroadcast
 from prime_rl.trainer.rl.broadcast.filesystem import FileSystemWeightBroadcast
 from prime_rl.trainer.rl.broadcast.nccl import NCCLWeightBroadcast
+from prime_rl.trainer.rl.broadcast.sparse_filesystem import SparseFileSystemWeightBroadcast
 
 
 def setup_weight_broadcast(
@@ -15,5 +16,7 @@ def setup_weight_broadcast(
         return NCCLWeightBroadcast(output_dir, config, torch.cuda.current_device())
     elif config.type == "filesystem":
         return FileSystemWeightBroadcast(output_dir, config, lora_config)
+    elif config.type == "sparse_filesystem":
+        return SparseFileSystemWeightBroadcast(output_dir, config, lora_config)
     else:
         raise ValueError(f"Invalid weight broadcast type: {config.type}")
